Rm the core project, get rid of def macros, simplify the project structure
