--- core/src/main/scala/ch/jodersky/jni/syntax/NativeLoader.scala
@@ -1,7 +1,0 @@
-package ch.jodersky.jni.syntax
-
-import ch.jodersky.jni.nativeLoaderMacro
-
-class NativeLoader(nativeLibrary: String) {
-  nativeLoaderMacro.load(nativeLibrary)
-}
