+++ macros/src/main/scala-3/ch/jodersky/jni/Process.scala
@@ -1,0 +1,10 @@
+package ch.jodersky.jni
+
+object Process {
+  def out(command: String): String =
+    try {
+      scala.sys.process.Process("uname -sm").lazyLines.head
+    } catch {
+      case ex: Exception => sys.error("Error running `uname` command")
+    }
+}
