+++ plugin/src/sbt-test/sbt-jni/simple-syntax/build.sbt
@@ -5,8 +5,8 @@ lazy val root = (project in file(".")).aggregate(core, native)
 lazy val core = project
   .settings(libraryDependencies += "org.scalatest" %% "scalatest" % "3.2.9" % Test)
   .settings(javah / target := (native / nativeCompile / sourceDirectory).value / "include")
+  .settings(macroProvided := false)
   .dependsOn(native % Runtime)
-  .enablePlugins(JniSyntax)
 
 lazy val native = project
   .settings(nativeCompile / sourceDirectory := sourceDirectory.value)
